Added implementations of Mov and Jmp

+++ src/LibEphedra/LibEphedra/VM/CPU.cs
@@ -10,7 +10,7 @@ namespace LibEphedra.VM
         private RAM ram;
         private List<Register> registers;
 
-        public Register GetRegister(byte identifier)
+        public Register GetRegister(UInt16 identifier)
         {
             if (identifier >= 0 && registers.Count < identifier)
                 return registers[identifier];
